Extract filesystem implementations to pkg/filesystem package
Moves local, archive (.zip/.tar/.rar/.7z), git, and nested filesystem
implementations from pkg/gowebserver to the new standalone pkg/filesystem
library. The only net/http coupling (newHandlerFromFS) remains in
gowebserver as a thin adapter. SanitizeFileName is now exported from
pkg/filesystem and used by upload.go.
